👷 (ci): Add Docker integration and Dependabot configuration
This commit introduces Docker files and settings, enabling the application to be built and run inside a Docker container. Additionally, a new Dependabot configuration file is implemented to automatically keep npm package dependencies up-to-date on a weekly basis. A GitHub workflow triggering Docker Image CI on push and pull request to the main branch is also included.